Backend/requirements/base.txt
appdirs==1.4.4
asgiref==3.3.4
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
cfgv==3.2.0
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coverage==5.5
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.1
dj-rest-auth==2.1.4
Django==3.2
django-admin-sortable==2.2.3
django-allauth==0.44.0
django-extensions==3.1.2
django-filter==2.4.0
django-model-utils==4.1.1
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.1.0
filelock==3.0.12
identify==2.2.3
idna==2.10
inflection==0.5.1
iniconfig==1.1.1
Markdown==3.3.4
mypy-extensions==0.4.3
nodeenv==1.6.0
oauthlib==3.1.0
packaging==20.9
pathspec==0.8.1
Pillow==8.2.0
pluggy==0.13.1
pre-commit==2.12.1
psycopg2==2.8.6
py==1.10.0
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
pytest-factoryboy==2.1.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
shortuuid==1.0.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
virtualenv==20.4.3
whitenoise==5.2.0
django-cors-headers==3.7.0
coreapi==2.3.3